test: add OpLoad fixtures for test managers
